refactor(step3): menu package로 이동
