+++ src/main/java/kitchenpos/application/OrderService.java
@@ -5,6 +5,8 @@ import kitchenpos.infra.KitchenridersClient;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderRepository;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
