+++ src/main/java/kitchenpos/domain/OrderLineItem.java
@@ -3,7 +3,7 @@ package kitchenpos.domain;
 import javax.persistence.*;
 import java.math.BigDecimal;
 import java.util.UUID;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 
 @Table(name = "order_line_item")
 @Entity
