+++ src/main/java/kitchenpos/menu/domain/JpaMenuRepository.java
@@ -1,11 +1,10 @@
-package kitchenpos.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
+package kitchenpos.menu.domain;
 
 import java.util.List;
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 public interface JpaMenuRepository extends MenuRepository, JpaRepository<Menu, UUID> {
     @Query("select m from Menu m, MenuProduct mp where mp.product.id = :productId")
