+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -1,9 +1,19 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
-import javax.persistence.*;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.ForeignKey;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import javax.persistence.Transient;
+import kitchenpos.domain.MenuGroup;
 
 @Table(name = "menu")
 @Entity
