+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -1,7 +1,16 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
-import javax.persistence.*;
 import java.util.UUID;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.ForeignKey;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+import javax.persistence.Transient;
 import kitchenpos.product.domain.Product;
 
 @Table(name = "menu_product")
