+++ src/main/java/kitchenpos/menu/ui/MenuRestController.java
@@ -1,13 +1,18 @@
-package kitchenpos.ui;
-
-import kitchenpos.application.MenuService;
-import kitchenpos.domain.Menu;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+package kitchenpos.menu.ui;
 
 import java.net.URI;
 import java.util.List;
 import java.util.UUID;
+import kitchenpos.menu.application.MenuService;
+import kitchenpos.menu.domain.Menu;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RequestMapping("/api/menus")
 @RestController
