+++ src/test/java/kitchenpos/Fixtures.java
@@ -9,6 +9,8 @@ import java.util.Random;
 import java.util.UUID;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.product.domain.Product;
 
 public class Fixtures {
